+++ tests/receipt.test.ts
@@ -17,7 +17,10 @@ it('issue & verify', async () => {
     }))
 
 
-  const privateKeyJwk = await cose.key.generate<JWK>('ES256', 'application/jwk+json')
+  const privateKeyJwk = await cose.crypto.key.gen<'ES256', 'application/jwk+json'>({
+    type: "application/jwk+json",
+    algorithm: "ES256"
+  })
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   const { d, ...publicKeyJwk } = privateKeyJwk
   const signer = cose.detached.signer({
